Several minor adjustments - most of them refer to formal matters

+++ j3l/src/j3l/util/sort/Sort.java
@@ -71,7 +71,7 @@ public final class Sort {
 							}
 						}
 						else {
-							return (int)(c1) - (int)(c2);
+							return c1 - c2;
 						}
 					}
